requirements.txt
pytest
pytest-cov
pytest-django
pytest-pep8
pytest-pythonpath
